Move exception to correct folder and abstract grouping logic
