+++ tests/Feature/Listeners/Workflow/SyncQueueTest.php
@@ -6,7 +6,7 @@ use Illuminate\Support\Facades\Bus;
 use Illuminate\Testing\Assert;
 use JobStatus\Enums\MessageType;
 use JobStatus\Enums\Status;
-use JobStatus\Exception\JobCancelledException;
+use JobStatus\Exceptions\JobCancelledException;
 use JobStatus\Models\JobBatch;
 use JobStatus\Models\JobStatus;
 use JobStatus\Tests\fakes\AssertBatch;
